move test to a different file, optimize tests
